pyproject.toml
[tool.poetry]
name = "Telegram File Uploader"
version = "0.1.0"
description = ""
authors = ["Alex <120180702+moveitaway@users.noreply.github.com>"]

[tool.poetry.dependencies]
python = "^3.10"
aiogram = "^2.23.1"
aiofile = "^3.8.1"
aioshutil = "^1.2"
shortuuid = "^1.0.11"
sqlalchemy = "^2.0.6"
alembic = "^1.10.2"
python-dotenv = "^1.0.0"
asyncpg = "^0.27.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
